new model admin
So mongoose allows for an object-oriented style for models, so I am
playing with the idea of making admin its own model which extends user,
and gave it methods removeUser, removeHouse, modifyRating.

Added admin page to routes, created a skeleton html for listUsers, which
just lists tenants and owners.  If you have a design in mind, send me an
email, otherwise I'll work one out based on the other jade files.